Crate amd

Modules

commands 🔒

graphql 🔒

ids 🔒

reaction_roles 🔒

scheduler 🔒

This module is a simple cron equivalent. It spawns threads for the [Task]s that need to be completed.

tasks 🔒

A trait to define a job that needs to be executed regularly, for example checking for status updates daily.

utils 🔒

Structs

Data

Functions

event_handler 🔒

main 🔒

setup_tracing 🔒

Type Aliases

Context

Error

ReloadHandle