amd

Module tasks

A trait to define a job that needs to be executed regularly, for example checking for status updates daily.

Modules

lab_attendance 🔒

status_update 🔒

Traits

Task

A Task is any job that needs to be executed on a regular basis. A task has a function Task::run_in that returns the time till the next [’Task::run`] is run.

Functions

get_tasks

Analogous to crate::commands::get_commands, every task that is defined must be included in the returned vector in order for it to be scheduled.