amd::tasks

Trait Task

pub trait Task: Send + Sync {
    // Required methods
    fn name(&self) -> &str;
    fn run_in(&self) -> Duration;
    fn run<'life0, 'async_trait>(
        &'life0 self,
        ctx: Context,
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

A Task is any job that needs to be executed on a regular basis. A task has a function Task::run_in that returns the time till the next [’Task::run`] is run.

Required Methods

fn name(&self) -> &str

fn run_in(&self) -> Duration

fn run<'life0, 'async_trait>( &'life0 self, ctx: Context, ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,